Add the capability to read items from a file
Nowadays, it is really difficult to configure this GitHub action, if you
have many events, logs, metrics, check to push to DataDog.
On the other hand, if the list of items is generated dynamically, then
you have to fight to generate a valid yaml string.

This commit adds the capability to configure a source yaml file to read
the items from.